Add folder structure to project
