+++ Assets/Art.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61bf005243a882449bb95c10845f661b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
