+++ Assets/Art/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c63bf826cabbeb5429404807bea4cdc0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
