+++ Assets/Art/Animations/UI/KnightIdleUI.anim.meta
