+++ Assets/Art/Animations/UI/KnightWalk_Clip.anim.meta
