+++ Assets/Art/Sprites.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 67ea0acc6717898459c0d0ca6d3c3cfa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
