+++ Assets/Art/Sprites/Soldiers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ec44d87e342968f489bb336ef1e174cf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
