+++ Assets/Art/Sprites/Soldiers/Knight.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5baff6b048aa80943bc6756a15bcdff2
+guid: 7b1d3b5dd6bae8744a37102a11448533
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
