+++ Assets/Art/Sprites/Soldiers/Knight/KnightIdle.aseprite
