+++ Assets/Art/Sprites/Soldiers/Knight/KnightWalk.aseprite
