+++ Assets/Scripts/Core/Architecture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 957cb74c70572a2459e7010856d4b37a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
