+++ Assets/Scripts/Core/Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eac5da641d371e94fa82d4bff5e6e3da
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
