+++ Assets/Scripts/Core/Managers/GameManager.cs.meta
