+++ Assets/Scripts/Core/Utilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b90bbfd3b248fc43b624456a102db0b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
