+++ Assets/Scripts/UI/Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ac5ed41b71a8aba4bbcab17d09a7aae5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
