+++ UserSettings/Layouts/default-6000.dwlt
@@ -365,7 +365,7 @@ MonoBehaviour:
   m_Pos:
     serializedVersion: 2
     x: 0
-    y: 418
+    y: 363
     width: 991
     height: 568
   m_SerializedDataModeController:
@@ -466,7 +466,7 @@ MonoBehaviour:
   m_Pos:
     serializedVersion: 2
     x: 0
-    y: 79
+    y: 24
     width: 991
     height: 313
   m_SerializedDataModeController:
@@ -1207,8 +1207,8 @@ MonoBehaviour:
     m_TextWithWhitespace: "Hierarchy\u200B"
   m_Pos:
     serializedVersion: 2
-    x: 992
-    y: 79
+    x: 1
+    y: 24
     width: 517
     height: 411
   m_SerializedDataModeController:
@@ -1273,8 +1273,8 @@ MonoBehaviour:
     m_TextWithWhitespace: "Project\u200B"
   m_Pos:
     serializedVersion: 2
-    x: 992
-    y: 516
+    x: 1
+    y: 461
     width: 517
     height: 470
   m_SerializedDataModeController:
@@ -1319,7 +1319,7 @@ MonoBehaviour:
     scrollPos: {x: 0, y: 0}
     m_SelectedIDs: 84ac0000
     m_LastClickedID: 44164
-    m_ExpandedIDs: 0000000082ac000084ac000086ac000088ac00008aac00008cac000000ca9a3b
+    m_ExpandedIDs: 000000008aac00008cac00008eac000090ac000092ac000094ac0000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -1348,7 +1348,7 @@ MonoBehaviour:
     scrollPos: {x: 0, y: 0}
     m_SelectedIDs: 
     m_LastClickedID: 0
-    m_ExpandedIDs: 0000000082ac000084ac000086ac000088ac00008aac00008cac0000
+    m_ExpandedIDs: 000000008aac00008cac00008eac000090ac000092ac000094ac0000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -1427,8 +1427,8 @@ MonoBehaviour:
     m_TextWithWhitespace: "Inspector\u200B"
   m_Pos:
     serializedVersion: 2
-    x: 1511
-    y: 79
+    x: 1512
+    y: 24
     width: 408
     height: 907
   m_SerializedDataModeController:
